// src/types/types.ts
export interface LastRealeases {
  href: string
  items: ItemAlbum[]
  limit: number
  next: string
  offset: number
  previous: null
  total: number
}

export interface SearchTracks {
  href: string
  items: ItemsTrack[]
  limit: number
  next: string
  offset: number
  previous: null
  total: number
}

export interface ItemsTrack {
  album: Album
  artists: Artist[]
  available_markets: string[]
  disc_number: number
  duration_ms: number
  explicit: boolean
  external_ids: ExternalIDS
  external_urls: ExternalUrls
  href: string
  id: string
  is_local: boolean
  name: string
  popularity: number
  preview_url: string
  track_number: number
  type: ItemType
  uri: string
}

export interface Album {
  album_type: AlbumTypeEnum
  artists: Artist[]
  available_markets: string[]
  external_urls: ExternalUrls
  href: string
  id: string
  images: Image[]
  name: string
  release_date: string
  release_date_precision: ReleaseDatePrecision
  total_tracks: number
  type: AlbumTypeEnum
  uri: string
}

export enum AlbumTypeEnum {
  Album = "album",
  Single = "single",
}

export interface Artist {
  external_urls: ExternalUrls
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls {
  spotify: string
}

export interface Image {
  height: number
  url: string
  width: number
}


export interface ExternalIDS {
  isrc: string
}

export enum ItemType {
  Track = "track",
}


export interface ItemAlbum {
  album_type: AlbumTypeEnum
  artists: Artist[]
  available_markets: string[]
  external_urls: ExternalUrls
  href: string
  id: string
  images: Image[]
  name: string
  release_date: string
  release_date_precision: ReleaseDatePrecision
  total_tracks: number
  type: AlbumTypeEnum
  uri: string
  popularity?: string
  tracks?: {
    items: ItemsTrack[]
  }
}


export interface Artist {
  external_urls: ExternalUrls
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls {
  spotify: string
}


export interface Image {
  height: number
  url: string
  width: number
}

export enum ReleaseDatePrecision {
  Day = "day",
}